refactor : 테이블 구조 변경 등
* 기존 1:1 관계로 분리되있었던 sidedish 테이블과 sidedish_price 테이블을 하나로 통합
* domain 클래스 내 Long 타입을 Integer 타입으로 변경

+++ BE/src/main/java/com/sidedish/dto/SideDishDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class SideDishDto {
 
-    private Long id;
+    private Integer id;
     private String name;
     private String description;
     private boolean dawnDeliveryFlag;
@@ -15,7 +15,7 @@ public class SideDishDto {
     private Integer discountRate;
     private String saveFileName;
 
-    public SideDishDto(Long id, String name, String description, boolean dawnDeliveryFlag,
+    public SideDishDto(Integer id, String name, String description, boolean dawnDeliveryFlag,
         boolean wholeNationDeliveryFlag, Integer price, String discountType,
         Integer discountRate, String saveFileName) {
         this.id = id;
